autoencoder logic and config out of main
